# openapi.yaml
openapi: 3.1.0
info:
  title: GoFile / fGo API (v0.1.0)
  version: 0.1.0
  summary: Minimal git‑lite file server API. Idempotent blob uploads, plan→finalize place, linear history on `main`.
  cmd_map:
    space: namespace
    box: repo
    arm: branch
    enact: commit
    dupe: clone
    grab: pull
    call: fetch
    place: push
    replace: force‑push to branch head
    store: push to specific arm
  description: |
    **Alpha notice**: pre‑release, breaking changes allowed until v1.0.
    fGo : Git Command Map:
      space ⇄ namespace
      box ⇄ repo
      arm ⇄ branch
      enact ⇄ commit
      dupe ⇄ clone
      grab ⇄ pull
      call ⇄ fetch
      place ⇄ push
      replace ⇄ force‑push to branch head
      store ⇄ push to specific arm

servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.gofiles.com
    description: Public Production server

tags:
  - name: System
  - name: Blobs
  - name: Spaces
  - name: Boxes
  - name: Arms
  - name: Enacts
  - name: Files

paths:
  /v1/health:
    get:
      tags: [ System ]
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  uptime: { type: number, format: float }

  /v1/blobs/{sha256}:
    parameters:
      - name: sha256
        in: path
        required: true
        description: Hex SHA‑256 digest of the blob
        schema:
          type: string
          pattern: '^[a-f0-9]{64}$'
    head:
      tags: [ Blobs ]
      summary: Check if blob exists (idempotent)
      security: [ ]
      responses:
        '200': { description: Present }
        '404': { description: Not found }
    put:
      tags: [ Blobs ]
      summary: Upload a blob by digest (idempotent)
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      parameters:
        - name: Digest
          in: header
          required: true
          description: RFC 3230 digest header `sha-256=BASE64`
          schema: { type: string }
        - name: Content-Length
          in: header
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '201': { description: Created }
        '204': { description: Already present }
        '400': { description: Bad digest/length }

  /v1/boxes:
    get:
      tags: [ Boxes ]
      summary: List boxes
      description: Public listing returns only `public` boxes in alpha.
      security: [ ]
      parameters:
        - name: visibility
          in: query
          schema:
            $ref: '#/components/schemas/Visibility'
      responses:
        '200':
          description: List of boxes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Box' }
    post:
      tags: [ Boxes ]
      summary: Create a new box
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name ]
              properties:
                name: { type: string, pattern: '^[a-z0-9-]{1,64}$' }
                visibility: { $ref: '#/components/schemas/Visibility' }
                default_arm: { type: string, default: main }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Box' } } } }
        '409': { description: Box already exists }

  /v1/boxes/{box}/place/plan:
    post:
      tags: [ Enacts ]
      summary: Plan a place (which blobs are missing)
      description: Computes missing digests for a manifest (paths→sha256,size).
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/box'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/placePlanRequest'
      responses:
        '200': { description: Plan result, content: { application/json: { schema: { $ref: '#/components/schemas/placePlanResponse' } } } }
        '400': { description: Invalid entries }

  /v1/boxes/{box}/place/finalize:
    post:
      tags: [ Enacts ]
      summary: Finalize a place and create a enact
      description: Verifies presence of all blobs, writes enact, moves arm head if parent matches.
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/box'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/placeFinalizeRequest'
      responses:
        '201': { description: Enact created, content: { application/json: { schema: { $ref: '#/components/schemas/placeFinalizeResponse' } } } }
        '409': { description: Parent mismatch / concurrent update }
        '422': { description: Digest/size mismatch }

  /v1/boxes/{box}/enacts/latest:
    get:
      tags: [ Enacts ]
      summary: Get latest enact for arm
      security: [ ]
      parameters:
        - $ref: '#/components/parameters/box'
        - in: query
          name: arm
          schema: { type: string, default: main }
      responses:
        '200': { description: Latest enact, content: { application/json: { schema: { $ref: '#/components/schemas/Enact' } } } }
        '404': { description: Not found }

  /v1/files/{enact_id}:
    get:
      tags: [ Files ]
      summary: Download a file by enact + path
      description: Use query `path` to specify file path inside the enact tree.
      security: [ ]
      parameters:
        - name: enact_id
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
        - name: Range
          in: header
          required: false
          schema: { type: string }
      responses:
        '200': { description: File content, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
        '206': { description: Partial content }
        '404': { description: Not found }

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque

  parameters:
    box:
      name: box
      in: path
      required: true
      description: Box name or ID
      schema: { type: string }

  schemas:
    Visibility:
      type: string
      enum: [ public, unlisted, private ]
      default: public

    Box:
      type: object
      required: [ id, name, visibility, default_arm ]
      properties:
        id: { type: string }
        name: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        default_arm: { type: string, example: main }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    Entry:
      type: object
      required: [ path, sha256, size, mode ]
      properties:
        path: { type: string, description: POSIX path within the box }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        size: { type: integer, minimum: 0 }
        mode: { type: integer, description: File mode (unix) }

    Enact:
      type: object
      required: [ id, box_id, arm, timestamp, entries ]
      properties:
        id: { type: string }
        box_id: { type: string }
        arm: { type: string }
        parent_id: { type: string, nullable: true }
        message: { type: string }
        author: { type: string }
        timestamp: { type: string, format: date-time }
        entries:
          type: array
          items: { $ref: '#/components/schemas/Entry' }

    placePlanRequest:
      type: object
      required: [ entries ]
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/Entry' }

    placePlanResponse:
      type: object
      required: [ missing, total ]
      properties:
        missing:
          type: array
          items:
            type: string
            pattern: '^[a-f0-9]{64}$'
        total: { type: integer, minimum: 0 }
        will_replace: { type: integer, minimum: 0 }

    placeFinalizeRequest:
      type: object
      required: [ arm, entries ]
      properties:
        arm: { type: string, default: main }
        parent_enact_id: { type: string, nullable: true }
        message: { type: string }
        entries:
          type: array
          items: { $ref: '#/components/schemas/Entry' }

    placeFinalizeResponse:
      type: object
      required: [ enact_id ]
      properties:
        enact_id: { type: string }
        uploaded: { type: integer }
        reused: { type: integer }

    Error:
      type: object
      required: [ error ]
      properties:
        error: { type: string }
        details: { type: object, additionalProperties: true }

security:
  - BearerAuth: [ ]